requirements.txt
pandas>=2.0.0
requests>=2.28.0
openpyxl>=3.1.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0
torch>=2.0.0
transformers>=4.20.0
